Change conditional for missions
