+++ src/components/pages/Missions.js
@@ -1,9 +1,14 @@
 import React from 'react';
-import JoinButton from '../conditionals/JoinButton';
+import { useDispatch } from 'react-redux';
+import { reserveMissions } from '../../Redux/State/MissionsState';
 import '../assets/styles/missions.css';
 
 const Missions = (missionsObj) => {
+  const dispatch = useDispatch();
   const { missions } = missionsObj;
+  const handleClick = (e) => {
+    dispatch(reserveMissions(e.target.id));
+  };
   const missionsIds = Object.keys(missions);
 
   return (
@@ -32,13 +37,20 @@ const Missions = (missionsObj) => {
                 </td>
                 <td>
                   <span
-                    className="notMember"
+                    className="member"
                   >
-                    Not a member
+                    {mission.reservation ? 'Active Member' : 'NOT A MEMBER'}
                   </span>
                 </td>
                 <td className="joinMission">
-                  <JoinButton missions={missionsObj} id={mission.information.mission_id} />
+                  <button
+                    className="joinMissionButton"
+                    id={mission.information.mission_id}
+                    onClick={(e) => handleClick(e)}
+                    type="button"
+                  >
+                    {mission.reservation ? 'Leave Mission' : 'Join Mission'}
+                  </button>
                 </td>
               </tr>
             ))}
